\name{fabryka}
\alias{fabryka}

\title{fabryka}

\description{Main package function to run the app.
}
\usage{fabryka()}

\value{
    Launch the `fabryka` R Shiny application.
}
\author{Marc Thomas <marcthomas1 at hotmail.fr>}


\examples{
    # launching the app:
    if(interactive()){
     fabryka()
    }
}
