\name{ternary_model}
\alias{ternary_model}
\docType{data}
\title{
Model data for Benn diagram
}
\description{
This dataset was sent by P. Bertran and contains data to build models describing the fabric of natural sedimentary processes.
}
\usage{data("ternary_model")}
\format{
  A data frame with 175 observations on the following 15 variables.
  \describe{
    \item{\code{Context}}{a character vector}
    \item{\code{Type}}{a character vector}
    \item{\code{Site}}{a character vector}
    \item{\code{Object}}{a character vector}
    \item{\code{n}}{a numeric vector}
    \item{\code{L}}{a numeric vector}
    \item{\code{p}}{a character vector}
    \item{\code{E1}}{a numeric vector}
    \item{\code{E2}}{a numeric vector}
    \item{\code{E3}}{a numeric vector}
    \item{\code{IS}}{a numeric vector}
    \item{\code{EL}}{a numeric vector}
    \item{\code{PL}}{a numeric vector}
    \item{\code{comments}}{a character vector}
    \item{\code{reference}}{a character vector}
  }
}
\source{
Bertran P., Bordes J.G., Barré A., Lenoble A., Mourre V. (2006) - Fabrique d’amas de débitage : données expérimentales. Bulletin de la Société Préhistorique Française 103, 33-47
Bertran P., Lenoble A. (2002) - Fabriques des niveaux archéologiques : Méthode et premier bilan des apports à l'étude taphonomique des sites paléolithiques. Paléo 14, pp. 13-28.
Bertran P., Hétu B., Texier J.P., Van Steijn H. (1997) - Fabric characteristics of slope deposits. Sedimentology 44, 1-16.
Lenoble A., Bertran P. (2004) - Fabric of Palaeolithic levels: methods and implications for site formation processes. Journal of Archaeological Science 31, 457-469.
}
\keyword{datasets}
